MOAR SONGS??? MOAR SONGS
I think there are like ... 26 songs in the game now? All of them will
need heavy revamping when it comes to stattributes (esp the interval
thing) but for the time being they seem to be okay balanced kinda maybe
I'll get on it after finals. Got to that point we talked about at the
meeting.